requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
paho-mqtt==1.6.1
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7
schedule==1.1.0
sqlparse==0.4.3
tzdata==2022.7
